requirements.txt
PyDispatcher
pyOpenSSL
service-identity
Twisted
